Fix inconsistency between Rust error expectations and python
The signature Result<Option<Block>, BlockStoreError> says that if there
isn't a block the value will be Ok(None) whereas the Python BlockStore
raises a Value error if a BlockWrapper can not be found for the
transaction id/batch id. This commit fixes the inconsistency by
converting a Python ValueError into Ok(None).

Signed-off-by: Boyd Johnson <bjohnson@bitwise.io>